+++ library/database/src/CHANGELOG.md
@@ -1,1 +1,9 @@
-Entity files are already up to date.
+# TypeORM entities
+
+Generated at 2026-06-21T17:10:36.156Z
+
+- CREATE `users` → library/database/src/entities/auth/User.ts
+- CREATE `sessions` → library/database/src/entities/auth/Session.ts
+- CREATE `accounts` → library/database/src/entities/auth/Account.ts
+- CREATE `verifications` → library/database/src/entities/auth/Verification.ts
+